+++ src/hooks/useAccounts.ts
@@ -3,7 +3,7 @@ import { Account } from "../types/Types";
 import { getAccounts } from "../Util/Supabase";
 import { useEffect, useState } from "react";
 
-export const useAccounts = () => {
+export const useAccounts = (accountKey: string = "") => {
     const [accounts, setAccounts] = useState<Account[]>([]);
     const [loading, setLoading] = useState(false);
     const [error, setError] = useState<string>("");
@@ -13,7 +13,11 @@ export const useAccounts = () => {
           setLoading(true);
           try {
             const response = await getAccounts(supabase);
-            setAccounts(response);
+            if(accountKey !== "") {
+                setAccounts(response.filter(account => account.key === accountKey));
+            } else {
+                setAccounts(response);
+            }
           } catch (error) {
             setError(error as string)
           } finally {
@@ -23,7 +27,7 @@ export const useAccounts = () => {
     
         fetchData()
     
-      }, []);
+      }, [accountKey]);
 
       return { accounts, loading, error }
 }
